Agent: Move json parsing error handler to http_island_api_client.py
http_requests_facade.py doesn't handle response parsing, so it shouldn't define json deserialization error handler
